Implementanto tela de ediçao de dados de serviços :sparkles

+++ totalClean/Classes/Servico.cs
@@ -10,7 +10,7 @@ namespace totalClean
     {
         public int id { get; set; }
         public String nome { get; set; }
-        public float preco { get; set; }
+        public double preco { get; set; }
         public Boolean ativo { get; set; }
 
     }
